refactor serializer code
